+++ converter/map_desc/ways_clip.py
